Fix: Nm should not have been a double

+++ include/crpropa/magneticField/TD13Field.h
@@ -100,7 +100,7 @@ private:
   std::vector<double> Ak;
   std::vector<double> k;
 
-  double Nm;
+  int Nm;
 
   int avx_Nm;
   int align_offset;
